[fix] Fix tests?
